+++ flutter_appauth/CHANGELOG.md
@@ -3,6 +3,7 @@
 * **Breaking change** all methods have now been made to return non-nullable types
 * [Android] updated plugin to specify `Theme.AppCompat.Translucent.NoTitleBar` as the theme for the `RedirectUriReceiverActivity` from the AppAuth Android SDK. This is to fix a crash raised with issues [#362](https://github.com/MaikuB/flutter_appauth/issues/362) and [#515](https://github.com/MaikuB/flutter_appauth/issues/515)
 * [iOS][macOS] bumped AppAuth iOS dependency to 1.7.5
+* Added `FlutterAppAuthOAuthError` class that contains string constants representing OAuth 2.0 error codes defined by the [specification](https://datatracker.ietf.org/doc/html/rfc6749#section-5.2).
 * Updated API docs with more details
 * Updated readme with more details on essential knowledge and links to OAuth 2.0 specifications